refactor: 프로젝트 폴더 구조 재편 및 접근 제어자 수정
- Features/{Public, Private} 구분을 제거하고 기능별(Features/*) 플랫 구조로 변경
- MainView 및 Sidebar 의존성 및 컴포넌트 참조 경로 수정
- ConnectivityToggle, UserProfile 등 네이밍 통일 및 정리
