+++ TodoMate/Features/Main/Sidebar.swift
@@ -14,6 +14,7 @@ import TodoMateDomain
 struct Sidebar: View {
   // MARK: - Properties
 
+  @Environment(NavigationManager.self) private var naviManager
   @Environment(TodoBoardStore.self) private var todoStore
   @Environment(MemoStore.self) private var memoStore
 
@@ -24,41 +25,34 @@ struct Sidebar: View {
   @AppStorage(UserDefaultsKey.cachedUserGroupId.rawValue)
   private var cachedUserGroupId: String = ""
 
-  @Binding var selection: NavigationDestination?
-
   // MARK: - Computed Properties
 
   private var todayTodoCount: Int {
-    let calendar = Calendar.current
-    return todoStore.todos.count(where: { calendar.isDateInToday($0.date) })
+    todoStore.todos.count(where: { $0.date.isToday })
   }
 
   private var memoCount: Int {
     memoStore.memos.count
   }
 
+  private var hasGroup: Bool {
+    !cachedUserGroupId.isEmpty
+  }
+
+  private var groupName: String {
+    (cachedGroupName.isEmpty ? nil : cachedGroupName) ?? "그룹"
+  }
+
   // MARK: - Body
 
   var body: some View {
-    List(selection: $selection) {
+    List(selection: Bindable(naviManager).selection) {
       Section {
         NavigationLink(value: NavigationDestination.settings) {
           if cachedProfileName.isEmpty {
-            UserProfileView(
-              displayName: "TodoMate",
-              style: .compact,
-              avatar: {
-                AppImage(size: 32)
-              },
-            )
+            placeholderProfile
           } else {
-            UserProfileView(
-              displayName: cachedProfileName,
-              style: .compact,
-              avatar: {
-                DefaultAvatar(displayName: cachedProfileName, size: 32)
-              },
-            )
+            userProfile
           }
         }
         .accessibilityIdentifier("sidebar_profile")
@@ -67,24 +61,12 @@ struct Sidebar: View {
 
       Section {
         NavigationLink(value: NavigationDestination.todo) {
-          Label {
-            Text("Todo")
-          } icon: {
-            Image(systemName: "checkmark.circle.fill")
-              .foregroundStyle(DesignSystem.Colors.primary)
-          }
-          .badge(Text("\(todayTodoCount)").monospacedDigit())
+          todoLabel
         }
         .accessibilityIdentifier("sidebar_todo")
 
         NavigationLink(value: NavigationDestination.memo) {
-          Label {
-            Text("Memo")
-          } icon: {
-            Image(systemName: "square.text.square.fill")
-              .foregroundStyle(DesignSystem.Colors.trafficYellow)
-          }
-          .badge(Text("\(memoCount)").monospacedDigit())
+          memoLabel
         }
         .accessibilityIdentifier("sidebar_memo")
       } header: {
@@ -104,7 +86,7 @@ struct Sidebar: View {
         HStack {
           Text("Public")
           Spacer()
-          PublicConnectivityToggle()
+          ConnectivityToggle()
             .padding(.trailing, 8)
         }
       }
@@ -114,6 +96,26 @@ struct Sidebar: View {
   }
 
   // MARK: - Subviews
+
+  private var todoLabel: some View {
+    Label {
+      Text("Todo")
+    } icon: {
+      Image(systemName: "checkmark.circle.fill")
+        .foregroundStyle(DesignSystem.Colors.primary)
+    }
+    .badge(Text("\(todayTodoCount)").monospacedDigit())
+  }
+
+  private var memoLabel: some View {
+    Label {
+      Text("Memo")
+    } icon: {
+      Image(systemName: "square.text.square.fill")
+        .foregroundStyle(DesignSystem.Colors.trafficYellow)
+    }
+    .badge(Text("\(memoCount)").monospacedDigit())
+  }
 
   private var groupLabel: some View {
     Label {
@@ -135,20 +137,33 @@ struct Sidebar: View {
     }
   }
 
-  private var hasGroup: Bool {
-    !cachedUserGroupId.isEmpty
+  private var placeholderProfile: some View {
+    UserProfile(
+      displayName: "TodoMate",
+      style: .compact,
+      avatar: {
+        AppImage(size: 32)
+      },
+    )
   }
 
-  private var groupName: String {
-    (cachedGroupName.isEmpty ? nil : cachedGroupName) ?? "그룹"
+  private var userProfile: some View {
+    UserProfile(
+      displayName: cachedProfileName,
+      style: .compact,
+      avatar: {
+        DefaultAvatar(displayName: cachedProfileName, size: 32)
+      },
+    )
   }
 }
 
 #Preview {
   NavigationSplitView {
-    Sidebar(selection: .constant(.todo))
+    Sidebar()
       .environment(TodoBoardStore.preview)
       .environment(MemoStore.preview)
+      .environment(NavigationManager.preview)
   } detail: {
     Text("Detail")
   }
